+++ example/src/main/scala/example/HelloWorld.scala
@@ -1,30 +1,16 @@
-package example
-
 import zhttp.http._
 import zhttp.service.Server
-import zio.stream.ZStream
-import zio.{App, ExitCode, URIO}
+import zio._
 
 object HelloWorld extends App {
 
-  def h1 = HttpApp.collectM { case req @ Method.POST -> !! / "foo" =>
-    req.decodeContent(ContentDecoder.text).map { content =>
-      Response(data = HttpData.fromText(content))
-    }
+  // Create HTTP route
+  val app: HttpApp[Any, Nothing] = HttpApp.collect {
+    case Method.GET -> !! / "text" => Response.text("Hello World!")
+    case Method.GET -> !! / "json" => Response.jsonString("""{"greetings": "Hello World!"}""")
   }
-
-  def h2 = HttpApp.collectM { case req @ Method.POST -> !! / "bar" =>
-    req.decodeContent(ContentDecoder.backPressure).map { content =>
-      req.getContentLength match {
-        case Some(value) => Response(data = HttpData.fromStream(ZStream.fromChunkQueue(content).take(value)))
-        case None        => Response.fromHttpError(HttpError.LengthRequired())
-      }
-    }
-  }
-
-  def app: HttpApp[Any, Throwable] = h1 +++ h2
 
   // Run it like any simple app
   override def run(args: List[String]): URIO[zio.ZEnv, ExitCode] =
-    Server.start(8090, app).exitCode
+    Server.start(8090, app.silent).exitCode
 }
